requirements.txt
requests == 2.32.5
psycopg2 == 2.9.11
watchdog == 6.0.0
rarfile == 4.2
py7zr == 1.1.2
simple-log-factory == 1.0.0
simple-log-factory-ext-otel == 1.4.0rc2
python-dotenv == 1.2.1
pillow == 12.1.1
paho-mqtt == 2.1.0
raccoontools == 1.2.1